Move ForkedDaapdUpdater to separate module
